requirements.txt
Pillow>=10.0
numpy>=1.30
pytest>=8.0
